feat: finish oddeven sort

+++ src/mpi/sort/oddeven.cpp
@@ -143,9 +143,8 @@ int main(int argc, char** argv)
         int partner_rank;
         char file_path[128];
 
-        int rx_cnt;
-        int tx_cnt;
-        int file_size_befor_merge; // save for merge result filtration
+        int rx_cnt = 0;
+        int tx_cnt = 0;
         std::vector<int> rx_buf(buf_size);
         std::vector<int> tx_buf(buf_size);
 
@@ -162,7 +161,7 @@ int main(int argc, char** argv)
                 else                partner_rank = world_rank + 1;
             }
             if (partner_rank < 0 || partner_rank == world_size) continue; // idle pass, no partner
-            printf("[phase %d node%d] %d<->%d\n", phase, world_rank, world_rank, partner_rank);
+            // printf("[phase %d node%d] %d<->%d\n", phase, world_rank, world_rank, partner_rank);
 
             // preparation for MPI_Sendrecv
             // input file
@@ -173,7 +172,6 @@ int main(int argc, char** argv)
                 fprintf(stderr, "node%d failed to open %s during phase %d\n", world_rank, file_path, phase);
                 MPI_Abort(MPI_COMM_WORLD, 1);
             }
-            file_size_befor_merge = std::filesystem::file_size(file_path);
             // output file
             sprintf(file_path, "data/mpi/node%d/sorted_partner.bin", world_rank);
             std::filesystem::create_directories(std::filesystem::path(std::string(file_path)).parent_path());
@@ -185,6 +183,9 @@ int main(int argc, char** argv)
             }
 
 
+            // record the send and recv data amount
+            int tx_ttl = 0;
+            int rx_ttl = 0;
             // start data exchange
             MPI_Status status;
             do {
@@ -192,6 +193,7 @@ int main(int argc, char** argv)
                 {
                     finput.read(reinterpret_cast<char*>(tx_buf.data()), sizeof(int) * buf_size);
                     tx_cnt = finput.gcount() / sizeof(int);
+                    tx_ttl += tx_cnt;
                     if (tx_cnt == 0) finput.close();
                 }
                 MPI_Sendrecv(
@@ -200,11 +202,10 @@ int main(int argc, char** argv)
                     MPI_COMM_WORLD, &status
                 );
                 MPI_Get_count(&status, MPI_INT, &rx_cnt);
+                rx_ttl += rx_cnt;
                 foutput.write(reinterpret_cast<char*>(rx_buf.data()), sizeof(int) * rx_cnt);
             } while (tx_cnt == buf_size || rx_cnt == buf_size);
             foutput.close();
-
-
 
             // start external merge
             std::vector<std::string> input_file_list;
@@ -218,20 +219,29 @@ int main(int argc, char** argv)
             sprintf(file_path, "data/mpi/node%d/merge.bin", world_rank);
             std::string merge_file_path = std::string(file_path);
             kmerge_file(input_file_list, merge_file_path);
-            std::filesystem::rename(merge_file_path, input_file_path1); // rename to "sorted.bin"
+            std::filesystem::rename(merge_file_path, input_file_path1); // replace the original "sorted.bin"
 
             // truncate corresponding part of each node
             {
+                // tx_ttl represents node's segment size
+                // rx_ttl represents partner's segment size
                 if (world_rank < partner_rank)
                 {
                     // keep the smaller part
-                    
+                    c_truncate(input_file_path1.c_str(), sizeof(int), 0     , tx_ttl, buf_size);
                 }
                 else
                 {
-                    // keep the bigger part
+                    // keep the larger part
+                    c_truncate(input_file_path1.c_str(), sizeof(int), rx_ttl, tx_ttl, buf_size);
                 }
             }
+
+            // printf(
+            //     "[node%d phase%d (%d<->%d)]: rx %d, tx %d, truncate %ld\n",
+            //     world_rank, phase, world_rank, partner_rank,
+            //     rx_ttl, tx_ttl, std::filesystem::file_size(input_file_path1) / sizeof(int)
+            // );
         }
 
     }
@@ -240,41 +250,50 @@ int main(int argc, char** argv)
 
     if (world_rank == 0)
     {
+        char file_path[128];
+        std::vector<int> in_buf(buf_size);
+
+        std::filesystem::create_directories("data/output");
+        std::ofstream foutput("data/output/final.bin", std::ofstream::binary);
         // only need one process to merge all sorted segments
         for (int i = 0; i < world_size; ++i)
         {
-            
+            sprintf(file_path, "data/mpi/node%d/sorted.bin", i);
+            std::ifstream finput(file_path, std::ifstream::binary);
+            if (!finput.is_open())
+            {
+                fprintf(stderr, "node0 failed to open reduce seg %s\n", file_path);
+                exit(1);
+            }
+
+            int in_cnt = 0;
+            int in_ttl = 0;
+            do {
+                finput.read(reinterpret_cast<char*>(in_buf.data()), sizeof(int) * buf_size);
+                in_cnt = finput.gcount() / sizeof(int);
+                in_ttl += in_cnt;
+                foutput.write(reinterpret_cast<char*>(in_buf.data()), sizeof(int) * in_cnt);
+            } while (in_cnt == buf_size);
+            // printf("%s %d\n", file_path, in_ttl);
+            finput.close();
         }
+        foutput.close();
     }
 
-    // put result to output folder
-    if (world_rank == 0)
+    if (delete_temp && world_rank == 0)
     {
-        char file_path[128];
-        sprintf(file_path, "data/mpi/node%d/sorted.bin", world_rank);
-        std::string result_path = std::string(file_path);
-        std::filesystem::create_directories(std::filesystem::path("data/output/final.bin").parent_path());
-        std::filesystem::rename(result_path, "data/output/final.bin");
-    }
-
-    if (delete_temp)
-    {
-        if (world_rank == 0)
+        try
         {
-            try
-            {
-                if (std::filesystem::exists("data/mpi"))
-                    std::filesystem::remove_all("data/mpi");
-                if (std::filesystem::exists("data/temp"))
-                    std::filesystem::remove_all("data/temp");
-            }
-            catch (const std::filesystem::filesystem_error& e)
-            {
-                std::cerr << "remove temporary files error: " << e.what() << std::endl;
-            }
+            if (std::filesystem::exists("data/mpi"))
+                std::filesystem::remove_all("data/mpi");
+            if (std::filesystem::exists("data/temp"))
+                std::filesystem::remove_all("data/temp");
+        }
+        catch (const std::filesystem::filesystem_error& e)
+        {
+            std::cerr << "remove temporary files error: " << e.what() << std::endl;
         }
     }
-
 
     MPI_Finalize();
     return 0;
